Bug-fix: hotkeys not working after changing window with keyboard
window.onkeyup was not firing if the user did something like press "Ctrl+Shift+N". This would leave Ctrl and Shift in the set of currently pressed keys, so all matches for correct hotkey presses were returning false. Added a listener for window.onblur that clears the set to fix the issue.